fix(OMN-3927): split models into one-per-file and use UUID for artifact_id
ONEX Architecture validator requires one model per file. Split
ModelSourceTrigger, ModelArtifactRegistryEntry, and ModelArtifactRegistry
into separate modules. Change artifact_id from str to UUID to satisfy
the Patterns validator.
